requirements.txt
requests
beautifulsoup4
openpyxl
gitpython 